backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
openai
langchain
langchain-core
langchain-community
langchain-openai
pgvector
python-dotenv
tiktoken
pydantic-settings
email-validator
